@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <UserSecretsId>aspnet-WatchIt.API-95040082-44C4-4E4F-9760-E3726CD1E980</UserSecretsId>
   </PropertyGroup>
 
@@ -14,7 +14,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\WatchIt.Application\WatchIt.Application.csproj" />
-    <ProjectReference Include="..\WatchIt.Domain\WatchIt.Domain.csproj" />
+    <ProjectReference Include="..\WatchIt.Domain\WatchIt.Data.csproj" />
   </ItemGroup>
 
 </Project>
