@@ -1,15 +1,9 @@
-﻿namespace WatchIt.Domain.Models
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
+﻿namespace WatchIt.Data.Models;
+using System.Collections.Generic;
 
-    public class Movie
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public IEnumerable<Like> Likes { get; set; }
-    }
+public class Movie
+{
+    public int Id { get; set; }
+    public string Name { get; set; }
+    public IEnumerable<Like> Likes { get; set; }
 }
