@@ -1,35 +1,35 @@
-﻿using System.Threading;
+﻿namespace WatchIt.Application.Commands.CreateMovie;
+
+using System.Threading;
 using System.Threading.Tasks;
 using AutoMapper;
 using MediatR;
 using WatchIt.Domain.Models;
 
-namespace WatchIt.Application.Commands.CreateMovie
+public class CreateMovieRequest : IRequest<CreateMovieResponse>
 {
-    public class CreateMovieRequest : IRequest<CreateMovieResponse>
+    public string Name { get; set; }
+    public string Description { get; set; }
+
+    public class CreateMovieHandler : IRequestHandler<CreateMovieRequest, CreateMovieResponse>
     {
-        public string Name { get; set; }
-        public string Description { get; set; }
+        private readonly IMapper mapper;
 
-        public class CreateMovieHandler : IRequestHandler<CreateMovieRequest, CreateMovieResponse>
+        public CreateMovieHandler(IMapper mapper)
         {
-            private readonly IMapper mapper;
+            this.mapper = mapper;
+        }
 
-            public CreateMovieHandler(IMapper mapper)
+        public async Task<CreateMovieResponse> Handle(CreateMovieRequest request, CancellationToken cancellationToken)
+        {
+            var movie = new Movie()
             {
-                this.mapper = mapper;
-            }
+                Id = 10
+            };
 
-            public async Task<CreateMovieResponse> Handle(CreateMovieRequest request, CancellationToken cancellationToken)
-            {
-                var movie = new Movie()
-                {
-                    Id = 10
-                };
-
-                var result = mapper.Map<CreateMovieResponse>(movie);
-                return await Task.FromResult(result);
-            }
+            var result = mapper.Map<CreateMovieResponse>(movie);
+            return await Task.FromResult(result);
         }
     }
 }
+
