@@ -1,10 +1,10 @@
-﻿using WatchIt.Application.Common;
+﻿namespace WatchIt.Application.Commands.CreateMovie;
+
+using WatchIt.Application.Common;
 using WatchIt.Domain.Models;
 
-namespace WatchIt.Application.Commands.CreateMovie
+public class CreateMovieResponse : IMapFrom<Movie>
 {
-    public class CreateMovieResponse : IMapFrom<Movie>
-    {
-        public int Id { get; set; }
-    }
+    public int Id { get; set; }
 }
+
