@@ -1,9 +1,8 @@
-﻿using AutoMapper;
+﻿
+namespace WatchIt.Application.Common;
 
-namespace WatchIt.Application.Common
-{
-    public interface IMapFrom<T> 
+public interface IMapFrom<T>
         where T : class
-    {
-    }
+{
 }
+
